Refactor fixtures for article_data route_handlers tests
